refactor: split dashboard app.js into 9 ES modules
Split 1553-line monolith into focused modules:
- ui-utils.js: error classification, clipboard, modals, icons
- auth.js: session state, cached API client singleton, theme, login
- billing.js: Stripe integration with payment integrity fix
- groups.js: multi-select builders, table rendering, CRUD
- keys.js: usage key table, 7-field permission modal, CRUD
- actions.js: IPFS action table, CRUD
- wallets.js: wallet table, CRUD
- runner.js: CodeJar editor, execute/getCid
- app.js: entry point orchestrating init

Bug fixes included:
- getClient() now caches singleton instead of new instance per call
- preloadAllTables uses Promise.allSettled with error banner
- Payment flow: separate confirmPayment catch for integrity
- Billing amount validation: minimum $5.00 (500 cents)
- Error classification: auth/network/server user-friendly messages
- Unified copyToClipboard utility replacing 3 inline patterns
- Button disable-on-click during async operations
- loadBillingBalance shows "Balance unavailable" on error
